requirements.txt
imageio==2.28.0 ; python_version >= '3.7'
networkx==3.1 ; python_version >= '3.8'  
numpy==1.24.3 ; python_version >= '3.8'  
packaging==23.1 ; python_version >= '3.7'
pillow==9.5.0 ; python_version >= '3.7'
pywavelets==1.4.1 ; python_version >= '3.8'
scikit-image==0.19.2
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
tifffile==2023.4.12 ; python_version >= '3.8'
django==4.1.7
pyrebase
urllib3
